test_webrtc/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
